baselines/datactive/requirements.txt
# torchvision~=0.14.1
matplotlib~=3.7.2
# sklearn~=0.0.post7
scikit-learn~=1.3.0
numpy~=1.23.0
pycocotools~=2.0.6
pillow~=9.4.0
lxml~=4.9.3
openpyxl~=3.1.2
scipy~=1.11.1
tqdm~=4.65.0
typing_extensions~=4.6.3
cleanlab~=2.5.0